messagin_systems/requirements.txt
amqp
celery
flask
celery[redis]
secure-smtplib
python-dotenv